requirements.txt
Pillow
numpy
opencv-python
sklearn
matplotlib
torch
torchvision
